% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialWeights.R
\name{bartlettSparse}
\alias{bartlettSparse}
\title{Weighting Kernel}
\usage{
bartlettSparse(d, dmax)
}
\arguments{
\item{d}{distance vector}

\item{dmax}{maximumm distance}
}
\value{
vector of bartlett weights
}
\description{
Weighting Kernel
}
