% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STsimFunctions.R
\name{df2stack}
\alias{df2stack}
\title{Convert list of dataframes with to rasterstack}
\usage{
df2stack(sim_i, DF)
}
\arguments{
\item{sim_i}{which simulation}

\item{DF}{list of dataframe}
}
\value{
rasterstack
}
\description{
Convert list of dataframes with to rasterstack
}
