% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STFakeData.R
\name{fake_data_traditional}
\alias{fake_data_traditional}
\title{Create Space Time Lattice Data}
\usage{
fake_data_traditional(n = 10, tf = 5, theta = c(5, 1, 1, 1))
}
\arguments{
\item{n}{spatial dimension}

\item{tf}{temporal dimension}

\item{theta}{parameter vector for RHS}
}
\value{
dataframe with (n1,n2,t) coordinates and variables Y,X,X, Country Time
}
\description{
Create Space Time Lattice Data
}
