% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpaceTimeSE_format.R
\name{is.lattice}
\alias{is.lattice}
\title{checks if data table has lattice structure ?}
\usage{
is.lattice(DAT, TIME, ID)
}
\arguments{
\item{DAT}{a data.table}

\item{TIME}{name of temporal column}

\item{ID}{name of cellular ID variable}
}
\description{
checks if data table has lattice structure ?
}
