% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcovST_meat.R
\name{iterateObsJSpatial}
\alias{iterateObsJSpatial}
\title{Wrapper for Matrix Calculation for Space}
\usage{
iterateObsJSpatial(sub_dat, Xvars, wmat, verbose = TRUE,
  checkSym = FALSE, XOmegaX = XOmegaX0, ...)
}
\arguments{
\item{sub_dat}{a dataframe object for 1 time period}

\item{Xvars}{RHS design matrix}

\item{wmat}{Weights Matrix}

\item{verbose}{print output}

\item{checkSym}{check if wmat is symmetric}

\item{XOmegaX}{meat function}
}
\value{
object to be used in vcov* functions
}
\description{
Wrapper for Matrix Calculation for Space
}
