% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialWeights.R
\name{KNN}
\alias{KNN}
\title{K nearest neighbours
Calculate the number of neighbours within a neighbourhood.}
\usage{
KNN(w, h = w, type = "Moore")
}
\arguments{
\item{w}{number of neighbours wide (east-west).}

\item{h}{number of neighbours long (north-south).}

\item{type}{type of neighbourhood; "Moore" or "VonNeumann"}
}
\value{
the number of nearest neighbours
}
\description{
K nearest neighbours
Calculate the number of neighbours within a neighbourhood.
}
\examples{
KNN( 4 )
}
