% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpaceTimeSE_format.R
\name{listj}
\alias{listj}
\title{Data Matrix Preparations}
\usage{
listj(ddff, wmat = NA, tmat = NA, t_cutoff = 4, rho_t = NA,
  d_cutoff = 1, rho_sp = NA, latlon = NA, convert_to_angles = TRUE)
}
\arguments{
\item{ddff}{formatted data.table from vcovST.format}

\item{wmat}{spatial weights matrix}

\item{tmat}{temporal weights matrix}

\item{t_cutoff}{temporal cutoff}

\item{rho_t}{unsupported vonneumann structure}

\item{d_cutoff}{distance cutoff}

\item{rho_sp}{unsupported vonneumann structure}

\item{latlon}{coordinates in lon,lat or x,y}

\item{convert_to_angles}{lon,lat to x,y?}
}
\value{
list object to be passed to vcov* functions
}
\description{
Data Matrix Preparations
}
