% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeData.R
\name{make_space_data}
\alias{make_space_data}
\alias{make_space_data.raster}
\title{Create Multivariate Spatial Data}
\usage{
make_space_data(
  dimS,
  K,
  t_id = NA,
  space_groups = round(sqrt(dimS)),
  xpars = c(0, 1),
  theta = 1:K,
  error_type = "distance",
  error_scale = 1,
  sar_factor = 0.5,
  wf_mat = rbind(rep(sar_factor, 3), c(sar_factor, 0, sar_factor), rep(sar_factor, 3)),
  xy_mat = expand.grid(Loc_X = seq(0, 1, length.out = dimS), Loc_Y = seq(1, 0,
    length.out = dimS)),
  ws_mat = exp(-as.matrix(dist(xy_mat)))
)

make_space_data.raster(
  dimS,
  K,
  t_id = NA,
  space_groups = round(sqrt(dimS)),
  xpars = c(0, 1),
  theta = 1:K,
  error_type = "distance",
  error_scale = 1,
  sar_factor = 0.5,
  wf_mat = rbind(rep(sar_factor, 3), c(sar_factor, 0, sar_factor), rep(sar_factor, 3)),
  xy_mat = expand.grid(Loc_X = seq(0, 1, length.out = dimS), Loc_Y = seq(1, 0,
    length.out = dimS)),
  ws_mat = exp(-as.matrix(dist(xy_mat)))
)
}
\arguments{
\item{dimS}{spatial dimension}

\item{K}{number of covariates}

\item{t_id}{Time ID}

\item{space_groups}{number of contiguous groups (e.g. countries)}

\item{xpars}{generate each Xvariable from uniform(xpar\link{1},xpar\link{2})}

\item{theta}{parameter vector for Y=X\%*\%theta}

\item{error_type}{c('distance', 'focal', 'spherical')}

\item{error_scale}{variance of errors (normally distributed)}

\item{sar_factor}{parameter for error_type='focal'}

\item{wf_mat}{matrix specifing weights to smooth when error_type='focal'}

\item{xy_mat}{matrix specifing xy locations}

\item{ws_mat}{matrix specifing correlation structer of error_type='distance'}
}
\value{
dataframe
}
\description{
Create Multivariate Spatial Data
}
\details{
'focal' creates errors based on \eqn{e = u + \rho v}.
'distance' (the default) creates errors base on drawing e from a multivariate normal with covariance matrix ws_mat.
}
\section{Functions}{
\itemize{
\item \code{make_space_data.raster}: make_space_data.df uses raster functions instead of matrices (primarily for transparent debugging)
}}

\examples{
library(STrollR)
DFs  <- make_space_data.raster(10,2,error_type='distance')
DFs  <- make_space_data.raster(10,2,error_type='focal')
DFs  <- make_space_data.raster(10,2,error_type='spherical')
DFs2 <- make_space_data(10,2,error_type='distance')
DFs2 <- make_space_data(10,2,error_type='focal')
DFs2 <- make_space_data(10,2,error_type='spherical')
}
