% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeData.R
\name{make_spacetime_data}
\alias{make_spacetime_data}
\title{Create Multivariate SpaceTime Data}
\usage{
make_spacetime_data(
  dimS,
  dimT,
  K,
  xpars = c(0, 1),
  theta = 1:K,
  space_groups = round(sqrt(dimS)),
  time_groups = round(sqrt(dimT)),
  error_type = "distance",
  error_scale = 1,
  xy_mat = expand.grid(Loc_X = seq(0, 1, length.out = dimS), Loc_Y = seq(1, 0,
    length.out = dimS)),
  ws_mat = exp(-1 * as.matrix(dist(xy_mat))),
  wt_mat = toeplitz((2/3)^(0:(dimT - 1))),
  wst_mat = kronecker(wt_mat, ws_mat),
  verbose = F
)
}
\arguments{
\item{dimS, }{dimT spatial dimension and temporal dimension}

\item{K}{number of covariates}

\item{xpars}{generate each Xvariable from uniform(xpar\link{1},xpar\link{2})}

\item{theta}{parameter vector for Y=X\%*\%theta}

\item{space_groups, }{time_groups number of contiguous groups (e.g. countries,  time-regimes)}

\item{error_type}{c('focal', 'distance')}

\item{error_scale}{variance of errors (normally distributed)}

\item{xy_mat}{matrix specifing xy locations}

\item{ws_mat}{matrix specifing correlation structer of error_type='distance'}

\item{wt_mat}{matrix specifing correlation structer of error_type='distance'}

\item{wst_mat=kronecker(wt_mat, }{ws_mat)}
}
\value{
dataframe
}
\description{
Create Multivariate SpaceTime Data
}
\details{
'focal' creates errors based on \eqn{e = u + \rho v}
'distance' creates errors base on drawing e from a multivariate normal with covariance matrix wst_mat
}
\examples{
DFst <- make_spacetime_data(11,6,2)
DFst[DFst$Space_ID==1,]
 
}
