% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeData.R
\name{make_time_data}
\alias{make_time_data}
\title{Create Multivariate Temporal Data}
\usage{
make_time_data(
  dimT,
  K,
  s_id = NA,
  time_groups = round(sqrt(dimT)),
  xpars = c(0, 1),
  theta = 1:K,
  error_type = "focal",
  error_scale = 1,
  ar_factor = 2/3,
  wt_mat = toeplitz(c(1, ar_factor, ar_factor^2, rep(0, dimT - 3))),
  xy_mat = data.frame(Loc_X = NA, Loc_Y = NA)
)
}
\arguments{
\item{dimT}{temporal dimension}

\item{K}{number of covariates}

\item{s_id}{Spatial ID}

\item{time_groups}{number of contiguous groups (e.g. time-regimes)}

\item{xpars}{generate each Xvariable from uniform(xpar\link{1},xpar\link{2})}

\item{theta}{parameter vector for Y=X\%*\%theta}

\item{error_type}{c('focal', 'distance', 'spherical')}

\item{error_scale}{variance of errors (normally distributed)}

\item{ar_factor}{parameter for error_type='focal'}

\item{wt_mat}{matrix specifing correlation structer of error_type='distance'}

\item{xy_mat}{matrix specifing xy locations (defaults to NA)}
}
\value{
dataframe
}
\description{
Create Multivariate Temporal Data
}
\details{
'focal' creates errors based on \eqn{e = u + \rho v}
'distance' creates errors base on drawing e from a multivariate normal with covariance matrix wt_mat
}
\examples{
DFt  <- make_time_data(6,2)

}
