% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfxall.R
\name{mfxall}
\alias{mfxall}
\title{Run multiple regressions on the same dataset}
\usage{
mfxall(FORMS, ..., mfx_fun = mfxi, mc.cores = as.numeric(system("nproc",
  intern = TRUE)), parallel = TRUE)
}
\arguments{
\item{FORMS}{list of regression formula}

\item{...}{args passed to mfx_fun}

\item{mfx_fun}{what type of regression}

\item{mc.cores}{number of cores if parallel=TRUE}

\item{parallel}{use parallel processing?}
}
\value{
List of Regressions Summaries
}
\description{
Run multiple regressions on the same dataset
}
