% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfxi_lm.R
\name{mfxi.lm}
\alias{mfxi.lm}
\title{Run a Regression}
\usage{
mfxi.lm(formi, datai, vcv = "standard")
}
\arguments{
\item{formi}{regression formula}

\item{datai}{data for regression}

\item{vcv}{type of covariance correction}
}
\value{
summary table
}
\description{
Run a Regression
}
