% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfxi.R
\name{mfxi}
\alias{mfxi}
\title{Run a Regression}
\usage{
mfxi(formi, datai, scl = TRUE, wmat0 = get("WMAT0"), unit_id = "ID",
  time_id = "Year", coord_id = c("x", "y"))
}
\arguments{
\item{formi}{regression formula}

\item{datai}{data for regression}

\item{scl}{vcov correct for spatial+temporal covariance?}

\item{wmat0}{spatial weights matrix passed to vcovSCL}

\item{unit_id, time_id, coord_id}{passed to vcovSCL}
}
\value{
summary table
}
\description{
Run a Regression
}
