% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STsimPlotFunctions.R
\name{DynPlot}
\alias{DynPlot}
\alias{DynGif}
\alias{mkGif}
\title{Create Gif Plots}
\usage{
DynPlot(DFlist, ti, pname = "STvarX", ind = 1)

DynGif(pname, vw = FALSE)

mkGif(DFlist, ti, pname = "STvarX", ind = 1, vw = FALSE)
}
\arguments{
\item{DFlist}{}

\item{ti}{number of time periods}

\item{pname}{name of file}

\item{ind}{which simulation}

\item{vw}{view output}
}
\value{
list of rasterstacks
}
\description{
Create Gif Plots
}
