% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STsimFunctions.R
\name{sim2stack}
\alias{sim2stack}
\title{Convert simulation to rasterstack}
\usage{
sim2stack(e_spt, nsim, xyt)
}
\arguments{
\item{e_spt}{matrix of draws from spam.mvtnorm  (each row a realization of a simulation)}

\item{xyt}{lattice structure}

\item{number}{of simulations}
}
\value{
list of rasterstacks
}
\description{
Convert simulation to rasterstack
}
