% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STsimFunctions.R
\name{spt_cleanup}
\alias{spt_cleanup}
\title{cleanup spam::mvtnorm}
\usage{
spt_cleanup(m, nsim, xyt)
}
\arguments{
\item{m}{matrix of simulations (each row a realization)}

\item{nsim}{number of simulations}

\item{xyt}{lattice coordinates}
}
\value{
dataframe
}
\description{
cleanup spam::mvtnorm
}
