% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STsimFunctions.R
\name{var2stack}
\alias{var2stack}
\title{Convert Dataframe with 1 variable to raster for one realization}
\usage{
var2stack(df_i, sim_i)
}
\arguments{
\item{df_i}{dataframe}

\item{sim_i}{which simulation}
}
\value{
raster
}
\description{
Convert Dataframe with 1 variable to raster for one realization
}
