% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarioJ.R
\name{varioJ}
\alias{varioJ}
\title{Variogram Calculation}
\usage{
varioJ(coords, cutt, residu, latlon = FALSE, indices = FALSE,
  clean = FALSE)
}
\arguments{
\item{coords}{coordinate matrix}

\item{cutt}{cutoff from which to calculate variogram}

\item{residu}{vector of values (i.e. OLS residuals) associated coords}

\item{latlon}{coordinates are lon,lat or x,y}

\item{indices}{return indices?}

\item{clean}{unused currently}
}
\value{
data.frame of dij and (ei-ej)^2
}
\description{
Variogram Calculation
}
