% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcovSHAC.R
\name{vcovSHAC}
\alias{vcovSHAC}
\alias{vcovSHACsep}
\alias{vcovSTHAC}
\title{Calculate a SHAC (Spatial Heteroskedastic and Autocorrelation Consistent) Variance Covariance Matrix}
\usage{
vcovSHAC(
  reg,
  wmat,
  method = "bruteforce_ll",
  cutoff_s,
  loc_lat,
  loc_lon,
  loc_y,
  loc_x,
  add_hc = T,
  add_cluster = F,
  add_hac = F,
  verbose = FALSE,
  manual_dist = T,
  ...
)

vcovSHACsep(
  reg,
  wmat,
  method = "bruteforce_llt",
  cutoff_s,
  cutoff_t,
  loc_lat,
  loc_lon,
  loc_y,
  loc_x,
  loc_t,
  add_hc = T,
  add_cluster = F,
  add_hac = F,
  verbose = FALSE,
  manual_dist = T,
  ...
)

vcovSTHAC(
  reg,
  wmat,
  method = "bruteforce_llt",
  cutoff_s,
  cutoff_t,
  loc_lat,
  loc_lon,
  loc_y,
  loc_x,
  loc_t,
  add_hc = T,
  add_cluster = F,
  add_hac = F,
  verbose = FALSE,
  manual_dist = T,
  ...
)
}
\arguments{
\item{reg}{an `lm' object}

\item{wmat}{weights matrix}

\item{method}{c(\verb{rolled', }semirolled', 'bruteforce_ll', 'bruteforce_xy')}

\item{cutoff_s}{include weights up to cutoff_s. Required to be in km if method='bruteforce_ll' or in map-units if 'bruteforce_xy'.}

\item{loc_lat, loc_lon}{required if method='bruteforce_ll'}

\item{loc_y, loc_x}{required if method='bruteforce_xy'}

\item{add_hc, add_cluster, add_hac}{logical for adding HC correction (default=T), Cluster correction (default=F), HAC correction (default=F)}

\item{verbose}{show messages}

\item{options}{passed to sandwich}
}
\value{
covariance matrix
}
\description{
Calculate a SHAC (Spatial Heteroskedastic and Autocorrelation Consistent) Variance Covariance Matrix
}
\section{Functions}{
\itemize{
\item \code{vcovSHACsep}: similar to vcovSTHAC, but treats spatial autocorrelation seperately within each time period

\item \code{vcovSTHAC}: Space and Time HAC. Also weights the time-dimension according to bartlett kernel (i.e., weight= K(space)*K(time)) for bartlett-kernel K.
}}

