% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcovSTsep_loop.R
\name{vcovSpace.loop}
\alias{vcovSpace.loop}
\title{vcovSpace with Parallel Approach}
\usage{
vcovSpace.loop(DAT, LISTJ = NA, wmat = LISTJ$wmat, verbose = FALSE,
  cores = 4)
}
\arguments{
\item{DAT}{list of regression objects from vcov.format}

\item{LISTJ}{list of weighting objects from listj}

\item{wmat}{weights matrix}

\item{verbose}{show messages}

\item{cores}{number of cores in spatial loop}
}
\value{
covariance matrix w/ spatial correction
}
\description{
vcovSpace with Parallel Approach
}
