% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcovSTsep_loop.R
\name{vcovSpace.single}
\alias{vcovSpace.single}
\title{vcovSpace, Single Year Only}
\usage{
vcovSpace.single(DAT, LISTJ = NA, wmat = LISTJ$wmat, verbose = FALSE)
}
\arguments{
\item{DAT}{list of regression objects from vcov.format}

\item{LISTJ}{list of weighting objects from listj}

\item{wmat}{weights matrix}

\item{verbose}{show messages}
}
\value{
covariance matrix w/ spatial correction
}
\description{
vcovSpace, Single Year Only
}
