% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpaceTimeSE_format.R
\name{vcovST.format}
\alias{vcovST.format}
\title{Format Regression Output}
\usage{
vcovST.format(reg, unit, time, sp_coords, DF = NA, SPLIT_sp = FALSE,
  SPLIT_t = FALSE, SPLIT = FALSE)
}
\arguments{
\item{reg}{felm object to be shaped}

\item{unit}{string name for cellular_id variable}

\item{time}{string name for time variable}

\item{sp_coords}{string name for coordinate variables}

\item{DF}{data.table to format, <NULL> formatts}

\item{SPLIT_sp}{return list dataframe of DF for each cell}

\item{SPLIT_t}{return list dataframe of DF for each time}

\item{SPLIT}{SPLIT_sp=SPLIT_sp=TRUE}
}
\value{
object to be used in vcov* functions
}
\description{
Format Regression Output
}
