% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpaceTimeSE.R
\name{vcovST}
\alias{vcovST}
\alias{vcovST.loop}
\title{calculate vcovSTsep for felm object}
\usage{
vcovST(vcovfun = vcovSTsep, reg, DF, unit, time, sp_coords,
  t_cutoff = NA, d_cutoff = NA, wmat = NA, tmat = NA,
  latlon = TRUE, convert_to_angles = TRUE, verbose = FALSE,
  SPLIT_sp = FALSE, SPLIT_t = FALSE, SPLIT = FALSE)

vcovST.loop(vcovfun = vcovSpace.loop, reg, DF, unit, time, sp_coords,
  t_cutoff = NA, d_cutoff = NA, wmat = NA, tmat = NA,
  latlon = TRUE, convert_to_angles = TRUE, verbose = FALSE,
  cores = 4, vcvtime = TRUE)
}
\arguments{
\item{vcovfun}{function for which type of SHAC correction}

\item{reg}{an lm or felm object}

\item{DF}{data.frame with unit, time, sp_coords}

\item{unit}{cellular id}

\item{time}{time id}

\item{sp_coords}{coordinate id}

\item{t_cutoff}{cutoff for considering time correlation}

\item{d_cutoff}{cutoff for considering space correlation}

\item{wmat}{spatial weights matrix}

\item{tmat}{temporal weights matrix}

\item{latlon}{are coordinates in lon,lat or x,y form}

\item{convert_to_angles}{convert lon,lat to x,y}

\item{verbose}{show messages}

\item{SPLIT_sp, SPLIT_t, SPLIT}{see vcovST.format}

\item{cores}{number of cores to use}

\item{vcvtime}{add temporal clustering}
}
\value{
covariance matrix w/ SHAC correction
}
\description{
calculate vcovSTsep for felm object

vcovST for seperable covariances using split approach
}
\details{
if vcvtime==FALSE, wmat should have diagonal elements
}
\section{Functions}{
\itemize{
\item \code{vcovST.loop}: 
}}

