% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcovSTsep_loop.R
\name{vcovSTsep.loop}
\alias{vcovSTsep.loop}
\title{vcovSTsep with Parallel for Space and Time}
\usage{
vcovSTsep.loop(DAT, LISTJ = NA, wmat = LISTJ$wmat,
  t_double_count = FALSE, verbose = FALSE, return_each = FALSE,
  cores = 4, tcores = NA)
}
\arguments{
\item{DAT}{list of regression objects from vcov.format}

\item{LISTJ}{list of weighting objects from listj}

\item{wmat}{weights matrix}

\item{t_double_count}{double count time?}

\item{verbose}{show messages}

\item{return_each}{for debugging, return only spatial and only temporal covariances}

\item{cores}{number of cores in spatial loop}

\item{tcores}{number of cores in temporal loop}
}
\value{
covariance matrix w/ SHAC correction
}
\description{
vcovSTsep with Parallel for Space and Time
}
