% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcovSTsep.R
\name{vcovSTsep}
\alias{vcovSTsep}
\title{vcovSTsep}
\usage{
vcovSTsep(DAT, LISTJ = NA, t_double_count = FALSE, verbose = FALSE,
  return_each = FALSE)
}
\arguments{
\item{DAT}{list of regression objects from vcov.format}

\item{LISTJ}{list of weighting objects from listj}

\item{t_double_count}{double count time?}

\item{verbose}{print output}

\item{return_each}{for debugging, return only spatial and only temporal covariances}
}
\value{
covariance matrix w/ SHAC correction
}
\description{
vcovSTsep
}
