% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialWeights.R
\name{VonNeumann}
\alias{VonNeumann}
\title{Compute VonNeumann Neighbours}
\usage{
VonNeumann(coord_sp, directions = 4)
}
\arguments{
\item{coord_sp}{SpatialPoints object or coordinate-matrix}

\item{directions}{see adjacent}
}
\value{
sparse weights matrix
}
\description{
Compute VonNeumann Neighbours
}
