% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialWeights.R
\name{weight_mat}
\alias{weight_mat}
\title{Compute Sparse Spatial Weights Matrix}
\usage{
weight_mat(M, cutoff, latlon = NA, convert_to_angles = TRUE)
}
\arguments{
\item{M}{matrix of coordinates}

\item{cutoff}{use distances up cutoff}

\item{latlon}{are the rows (lat,lon) coordinates?}

\item{convert_to_angles}{convert cutoff from km to angles?}
}
\value{
the number of nearest neighbours
}
\description{
Compute Sparse Spatial Weights Matrix
}
\examples{
weight_mat(expand.grid( list(x=1:10, y=1:10)), cutoff=.5)
}
