% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialWeights.R
\name{write.listw2gwt}
\alias{write.listw2gwt}
\title{Write a listw object as a GWT file}
\usage{
write.listw2gwt(listw, dgwt_outfile = paste0(tempdir(), "dgwt.GWT"))
}
\arguments{
\item{listw}{returned from spdep::listw}

\item{dgwt_outfile}{name of file}
}
\value{
dgwt_outfile
}
\description{
Write a listw object as a GWT file
}
