% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcovSHAC.R
\name{XOmegaX}
\alias{XOmegaX}
\alias{XOmegaX_rolled}
\alias{XOmegaX_semirolled}
\alias{XOmegaX_bruteforce_ll}
\alias{XOmegaX_bruteforce_xy}
\alias{XOmegaX_bruteforce_llt}
\alias{XOmegaX_bruteforce_xyt}
\title{XOmegaX
Meat Matrix Calculations}
\usage{
XOmegaX_rolled(X, WMAT, E)

XOmegaX_semirolled(X, WMAT, E)

XOmegaX_bruteforce_ll(X, E, loc_lat, loc_lon, cutoff_s, manual_dist = T)

XOmegaX_bruteforce_xy(X, E, loc_y, loc_x, cutoff_s, manual_dist = T)

XOmegaX_bruteforce_llt(
  X,
  E,
  loc_lat,
  loc_lon,
  loc_t,
  cutoff_s,
  cutoff_t,
  manual_dist = T
)

XOmegaX_bruteforce_xyt(
  X,
  E,
  loc_y,
  loc_x,
  loc_t,
  cutoff_s,
  cutoff_t,
  manual_dist = T
)
}
\arguments{
\item{X}{design matrix}

\item{WMAT}{weighting matrix (preferably sparse sparse)}

\item{E}{vector of residuals}

\item{loc_lat, }{loc_lon spatial-coordinate vectors}

\item{cutoff_s}{how far to extend bartlett kernel in spatial dimention (in km)}

\item{loc_y, }{loc_x spatial-coordinate vectors}

\item{loc_t}{time-coordinate vector}

\item{cutoff_t}{how far to extend bartlett kernel in time dimension (in time units)}
}
\value{
object to be used in vcov* functions
}
\description{
XOmegaX
Meat Matrix Calculations
}
\details{
XOmegaX_rolled Fully is for Sparse-Matrix Meat-Matrix Calculation (Whole Matrix At Once)
XOmegaX_semirolled is for Sparse-Matrix Meat-Matrix Calculation (One Row At A Time)
XOmegaX_bruteforce_ll is for latlon data.
XOmegaX_bruteforce_llt is for latlon and time data.
XOmegaX_bruteforce_xy is for projected data.
XOmegaX_bruteforce_xyt is for projected and time data.
XOmegaX_bruteforce_*t weights the time-dimension according to bartlett kernel
}
