% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcovST_meat.R
\name{XOmegaX0}
\alias{XOmegaX0}
\title{Main `Meat` Matrix Calculation}
\usage{
XOmegaX0(X, WMAT, e)
}
\arguments{
\item{X}{design matrix}

\item{WMAT}{weighting matrix (preferably sparse sparse)}

\item{e}{vector of residuals}
}
\value{
object to be used in vcov* functions
}
\description{
Main `Meat` Matrix Calculation
}
