% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialWeights.R
\name{bartlettSparse}
\alias{bartlettSparse}
\title{Weighting Kernel}
\usage{
bartlettSparse(d, dmax)
}
\arguments{
\item{d}{returned from spdep::listw}

\item{dmax}{name of file}
}
\value{
bartlett weight
}
\description{
Weighting Kernel
}
