% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STsimFunctions.R
\name{mkGif}
\alias{mkGif}
\title{Create Gifs}
\usage{
mkGif(
  DFlist,
  ti,
  fdir = "~/Desktop/Packages/STrollR/STsim/",
  pname = "STvarX",
  ind = 1,
  vw = FALSE
)
}
\arguments{
\item{DFlist}{}

\item{ti}{number of time periods}

\item{fdir, pname}{directory and file}

\item{ind}{which simulation}

\item{vw}{view output}
}
\value{
list of rasterstacks
}
\description{
Create Gifs
}
