% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarioJ.R
\name{varioJ}
\alias{varioJ}
\title{Variogram Calculation}
\usage{
varioJ(
  coords,
  cuttoff,
  E,
  latlon = FALSE,
  indices = FALSE,
  clean = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{coords}{coordinate matrix}

\item{cuttoff}{cutoff from which to calculate variogram}

\item{E}{vector of values (i.e. OLS residuals) associated coords}

\item{latlon}{coordinates are lon,lat or x,y}

\item{indices}{return indices?}

\item{clean}{unused currently}

\item{verbose}{}
}
\value{
data.frame of dij and (ei-ej)^2
}
\description{
Variogram Calculation
}
