% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialWeights.R
\name{VonNeumann}
\alias{VonNeumann}
\title{Compute VonNeumann Neighbours}
\usage{
VonNeumann(coord_sp)
}
\arguments{
\item{coord_sp}{SpatialPoints object, matrix of coordinates}
}
\value{
sparse weights matrix
}
\description{
Compute VonNeumann Neighbours
}
